channeler: renamed external facing functions
